requirements.txt
opencv-python
matplotlib
torchvision
pillow==6.1
https://download.pytorch.org/whl/cu90/torch-1.1.0-cp36-cp36m-win_amd64.whl
